Don't try to predigest the DB connection URL using TUrl.
The TUrl class is useful for parsing the pieces of a URL, but it tries to
"rationalize" the url before returning the URL value with GetUrl().
Unfortunately, it is a little two aggressive, and doesn't have a complete
understanding of possible URL syntax, so it can mangle some of the
conversions.  The new connection code still uses TUrl to parse the URL, but
hands the explicit user input to the TSQLServer::Connect() method.