Add comments and clean up code formatting.
This begins the process of commenting the Dbi (it has almost none to begin
with).  It also starts cleaning up the line wrap and indentation to make
sure that the comments are readable (e.g. keep a fixed column width), and
that the code wraps at logical breaks in the statements.