Add checks that the column times are limited to standard SQL.
Most SQL engines define data types beyond those defined in the SQL
standards.  Those types are fine as long as we can program to a specific
engine, but will cause the tables to be non-portable.  The TDbiFieldType
class has been modified to complain about any types that are not part of
the standard.