Fixes and formats for TVldTimeStamp.
This adds some tweaks to make the code compile with fewer warnings. The
most notable change is that the internal second field is not Long64_t so
this class will work without modification for the next 600 billion years or
so.  It still depends on std::time_t for the internal conversions, so we
only get the benefit of the expanded field as operating systems recompile
with std::time_t defined as int64_t.